refactor: moved VolumeCutout into its own file
