+++ cloudvolume/__init__.py
@@ -1,5 +1,6 @@
 from .cloudvolume import CloudVolume, EmptyVolumeException, EmptyRequestException
 from .provenance import DataLayerProvenance
+from .volumecutout import VolumeCutout
 from .storage import Storage
 from .threaded_queue import ThreadedQueue
 from .connectionpools import ConnectionPool
